requirements.txt
tqdm>=4.66.1
tabulate>=0.9.0
ruamel.yaml>=0.18.5
defusedxml>=0.7.1
iso3166>=2.1.1
protobuf==3.20.1
open-simulation-interface @ git+https://github.com/OpenSimulationInterface/open-simulation-interface.git@v3.6.0
